rfkill.py: explicitly set buffering=0
This fixes an issue with newer kernel versions - we had
always documented the kernel API has being able to extend
the struct size, but using buffering in python makes that
impossible. Fix this by not using buffering.

This really only matters on read, but set it on write for
consistency as well.